Split the value parsers out of config.py into nab_project/values.py
